requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
alpaca-py==0.29.0
pandas==2.2.3
pandas-ta==0.3.14b
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
APScheduler==3.10.4
pytz==2024.2
